webapi/webapi-messaging-w3c-tests/messaging/MmsMessage_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>Messaging Test: MmsMessage_Basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/messaging/#idl-def-MmsMessager">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>
var testNumber = "5556";
var filter = {
  type: "mms",
};
var filterOption = {
  sortBy: "date",
  sortBy: "descending",
  limit: 3
};
var t1 = async_test("Check if the readonly attribute messageID of MmsMessage exists and type of string", {timeout: 2000});
var t2 = async_test("Check if the readonly attribute type of MmsMessage exists and type of string", {timeout: 2000});
var t3 = async_test("Check if the readonly attribute serviceID of MmsMessage exists and type of string", {timeout: 2000});
var t4 = async_test("Check if the readonly attribute from of MmsMessage exists and type of string", {timeout: 2000});
var t5 = async_test("Check if the readonly attribute timestamp of MmsMessage exists and type of Date", {timeout: 2000});
var t6 = async_test("Check if the readonly attribute read of MmsMessage exists and type of boolean", {timeout: 2000});
var t7 = async_test("Check if the readonly attribute to of MmsMessage exists and type of array", {timeout: 2000});
var t8 = async_test("Check if the readonly attribute expiry of MmsMessage exists and type of string", {timeout: 2000});
var t9 = async_test("Check if the readonly attribute cc of MmsMessage exists and type of array", {timeout: 2000});
var t10 = async_test("Check if the readonly attribute bcc of MmsMessage exists and type of array", {timeout: 2000});
var t11 = async_test("Check if the readonly attribute subject of MmsMessage exists and type of string", {timeout: 2000});
var t12 = async_test("Check if the readonly attribute smil of MmsMessage exists and type of  string", {timeout: 2000});
var t13 = async_test("Check if the readonly attribute attachments of MmsMessage exists and type of  array", {timeout: 2000});
var t14 = async_test("Check if the readonly attribute state of MmsMessage exists and type of  string", {timeout: 2000});
var t15 = async_test("Check if the readonly attribute deliveryInfo of MmsMessage exists and type of  string", {timeout: 2000});
var t16 = async_test("Check if the readonly attribute readReportRequested of MmsMessage exists and type of  boolean", {timeout: 2000});


t1.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t1.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("messageID" in mmsMessage, "The messageID attribute of MmsMessage exists");
        var dc = mmsMessage.messageID;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", " The type messageID of MmsMessage should be string");
        mmsMessage.messageID = dc + "test";
        assert_equals(mmsMessage.messageID, dc, "The value of messageID");
      });
      t1.done();
    },
    function (error) {
      t1.step(function () {
        assert_unreached(error.message);
      });
      t1.done();
    }
  );
});

t2.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t2.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("type" in mmsMessage, "The type attribute of MmsMessage exists");
        var dc = mmsMessage.type;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of mmsMessage.type should be string");
        mmsMessage.type = dc + "test";
        assert_equals(mmsMessage.type, dc, "The value of type");
      });
      t2.done();
    },
    function (error) {
      t2.step(function () {
        assert_unreached(error.message);
      });
      t2.done();
    }
  );
});

t3.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t3.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("serviceID" in mmsMessage, "The attribute of MmsMessage exists");
        var dc = mmsMessage.serviceID;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of mmsMessage.serviceID should be string");
        mmsMessage.serviceID = dc + "test";
        assert_equals(mmsMessage.serviceID, dc, "The value of  serviceID");
      });
      t3.done();
    },
    function (error) {
      t3.step(function () {
        assert_unreached(error.message);
      });
      t3.done();
    }
  );
});

t4.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t4.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("from" in mmsMessage, "The from attribute of MmsMessage exists");
        var dc = mmsMessage.from;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "type of mmsMessage.from should be string");
        mmsMessage.from = dc + "test";
        assert_equals(mmsMessage.from, dc, "The value of from");
      });
      t4.done();
    },
    function (error) {
      t4.step(function () {
        assert_unreached(error.message);
      });
      t4.done();
    }
  );
});

t5.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t5.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("timestamp" in mmsMessage, "The timestamp attribute of MmsMessage exists");
        var dc = mmsMessage.timestamp;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of mmsMessage.timestamp should be string");
        mmsMessage.timestamp = dc + "test";
        assert_equals(mmsMessage.serviceID, dc, "The value of timestamp");
      });
      t5.done();
    },
    function (error) {
      t5.step(function () {
        assert_unreached(error.message);
      });
      t5.done();
    }
  );
});

t6.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t6.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("read" in mmsMessage, "The read attribute of MmsMessage exists");
        var dc = mmsMessage.read;
        assert_equals(Object.prototype.toString.call(dc), "[object boolean]", "The type of mmsMessage.read should be boolean");
        if(dc) {
          mmsMessage.read = false;
        } else {
          mmsMessage.read = true;
        }
        assert_equals(mmsMessage.read, dc, "The value of read");
      });
      t6.done();
    },
    function (error) {
      t6.step(function () {
        assert_unreached(error.message);
      });
      t6.done();
    }
  );
});

t7.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t7.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("to" in mmsMessage, "The to attribute of MmsMessage exists");
        var dc = mmsMessage.to;
        assert_equals(Object.prototype.toString.call(dc), "[object Array]", "The type of mmsMessage.to should be string");
        mmsMessage.to = [];
        assert_equals(mmsMessage.to, dc, "The value of to");
      });
      t7.done();
    },
    function (error) {
      t7.step(function () {
        assert_unreached(error.message);
      });
      t7.done();
    }
  );
});

t8.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t8.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("expiry" in mmsMessage, "The expiry attribute of MmsMessage exists");
        var dc = mmsMessage.expiry;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of mmsMessage.expiry should be string");
        mmsMessage.body = dc + "test";
        assert_equals(mmsMessage.expiry, dc, "The value of expiry");
      });
      t8.done();
    },
    function (error) {
      t8.step(function () {
        assert_unreached(error.message);
      });
      t8.done();
    }
  );
});

t9.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t9.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("cc" in mmsMessage, "The cc attribute of MmsMessage exists");
        var dc = mmsMessage.cc;
        assert_equals(Object.prototype.toString.call(dc), "[object Array]", "The type of mmsMessage.cc should be Array");
        mmsMessage.cc = ["test"];
        assert_equals(mmsMessage.cc, dc, "The value of cc");
      });
      t9.done();
    },
    function (error) {
      t9.step(function () {
        assert_unreached(error.message);
      });
      t9.done();
    }
  );
});

t10.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t10.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("bcc" in mmsMessage, "The bcc attribute of MmsMessage exists");
        var dc = mmsMessage.bcc;
        assert_equals(Object.prototype.toString.call(dc), "[object Array]", "type of mmsMessage.bcc should be Array");
        mmsMessage.bcc = ["test"];
        assert_equals(mmsMessage.bcc, dc, "The value of bcc");
      });
      t10.done();
    },
    function (error) {
      t10.step(function () {
        assert_unreached(error.message);
      });
      t10.done();
    }
  );
});

t11.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t11.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("subject" in mmsMessage, "The subject attribute of MmsMessage exists");
        var dc = mmsMessage.subject;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of mmsMessage.subject should be string");
        mmsMessage.subject = dc + "test";
        assert_equals(mmsMessage.subject, dc, "The value of subject ");
      });
      t11.done();
    },
    function (error) {
      t11.step(function () {
        assert_unreached(error.message);
      });
      t11.done();
    }
  );
});

t12.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t12.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("smil" in mmsMessage, "The smil attribute of MmsMessage exists");
        var dc = mmsMessage.smil;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of mmsMessage.smil should be string");
        mmsMessage.smil = dc + "test";
        assert_equals(mmsMessage.smil, dc, "The smil attribute should be readonly");
      });
      t12.done();
    },
    function (error) {
      t12.step(function () {
        assert_unreached(error.message);
      });
      t12.done();
    }
  );
});

t13.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t13.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("attachments" in mmsMessage, "The attachments attribute of MmsMessage exists");
        var dc = mmsMessage.attachments;
        assert_equals(Object.prototype.toString.call(dc), "[object Array]", "The type of mmsMessage.attachments should be array");
        mmsMessage.attachments = ["test"];
        assert_equals(mmsMessage.attachments, dc, "The value of attachments");
      });
      t13.done();
    },
    function (error) {
      t13.step(function () {
        assert_unreached(error.message);
      });
      t13.done();
    }
  );
});

t14.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t14.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("state" in mmsMessage, "The state attribute of MmsMessage exists");
        var dc = mmsMessage.state;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of mmsMessage.state should be string");
        mmsMessage.state = dc + "test";
        assert_equals(mmsMessage.state, dc, "The value of state");
      });
      t14.done();
    },
    function (error) {
      t14.step(function () {
        assert_unreached(error.message);
      });
      t14.done();
    }
  );
});

t15.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t15.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("deliveryInfo" in mmsMessage, "The deliveryInfo attribute of MmsMessage exists");
        var dc = mmsMessage.deliveryInfo;
        assert_equals(Object.prototype.toString.call(dc), "[object Array]", "The type of mmsMessage.deliveryInfo should be array");
        mmsMessage.deliveryInfo = ["test"];
        assert_equals(mmsMessage.deliveryInfo, dc, "The value of deliveryInfo");
      });
      t15.done();
    },
    function (error) {
      t15.step(function () {
        assert_unreached(error.message);
      });
      t15.done();
    }
  );
});

t16.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t16.step(function () {
        assert_true(msg.element.length > 0, "no mms message");
        var mmsMessage = msg.element[0];
        assert_true("readReportRequested" in mmsMessage, "The readReportRequested attribute of MmsMessage exists");
        var dc = mmsMessage.readReportRequested;
        assert_equals(Object.prototype.toString.call(dc), "[object boolean]", "The type of mmsMessage.readReportRequested should be boolean");
        if(dc) {
          mmsMessage.readReportRequested = false;
        } else {
          mmsMessage.readReportRequested = true;
        }
        assert_equals(mmsMessage.readReportRequested, dc, "The value of readReportRequested");
      });
      t16.done();
    },
    function (error) {
      t16.step(function () {
        assert_unreached(error.message);
      });
      t16.done();
    }
  );
});
</script>
